# docs/swagger.yaml
basePath: /
definitions:
  handlers.errorDetail:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  handlers.errorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.errorDetail'
    type: object
  models.Payment:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
    type: object
  models.PaymentStatus:
    enum:
    - Authorized
    - Declined
    - Rejected
    type: string
    x-enum-varnames:
    - StatusAuthorized
    - StatusDeclined
    - StatusRejected
  models.PostPaymentRequest:
    properties:
      amount:
        type: integer
      card_number:
        type: string
      currency:
        type: string
      cvv:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
    type: object
host: localhost:8090
info:
  contact: {}
  description: Interview challenge for building a Payment Gateway - Go version
  title: Payment Gateway Challenge Go
paths:
  /api/payments:
    post:
      consumes:
      - application/json
      description: Submits a card payment to the acquiring bank and returns its outcome.
        An optional Idempotency-Key header makes a retried request replay the original
        outcome instead of being reprocessed.
      parameters:
      - description: Client-generated key that makes retrying this request safe
        in: header
        name: Idempotency-Key
        type: string
      - description: Payment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PostPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: Malformed request body
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "409":
          description: A request with this Idempotency-Key is already in progress
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "422":
          description: 'Rejected: one or more fields failed validation, or Idempotency-Key
            was reused with a different request body'
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "502":
          description: Bank unavailable
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Process a payment
      tags:
      - payments
  /api/payments/{id}:
    get:
      description: Retrieves the details of a previously processed payment by its
        ID.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Retrieve a payment
      tags:
      - payments
swagger: "2.0"